Changed Dictionaries to use ArrayLists
I have not modified any of the other files that may use Dictionaries. That NEEDS to be accounted for before this branch is merged into main.
If nobody fixes it by Friday I'll do it in the evening. It's just a result of me working from github desktop editor instead of intellij cause I didn't bring my laptop with me today :)